Move requests to its own package
