+++ client/package.json
@@ -33,6 +33,7 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "postcss": "^8.4.31",
     "prettier-plugin-tailwindcss": "^0.5.6",
+    "requests": "^1.0.0",
     "tailwindcss": "^3.3.5",
     "typescript": "^5.0.2",
     "vite": "^4.4.5",
