+++ requests/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "requests",
+  "version": "1.0.0",
+  "scripts": {
+    "build": "echo No build action"
+  },
+  "type": "module",
+  "types": "index.d.ts"
+}
