+++ server/package.json
@@ -27,6 +27,7 @@
     "database": "^1.0.0",
     "express": "^4.18.2",
     "http-proxy-middleware": "^2.0.6",
-    "mongoose": "^8.1.0"
+    "mongoose": "^8.1.0",
+    "requests": "^1.0.0"
   }
 }
