--- server/requests/package.json
@@ -1,11 +1,0 @@
-{
-    "name": "server-requests",
-    "version": "1.0.0",
-    "description": "",
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "type": "module",
-    "types": "index.d.ts"
-  }
-
