+++ server/src/addFakeData.ts
@@ -1,7 +1,7 @@
 import { startDockerContainer } from 'database';
 import mongoose from 'mongoose';
 import { matchApp } from './Schema.js';
-import { MatchData } from 'server/requests';
+import { MatchData } from 'requests';
 
 function randint(max: number, min = 0) {
     return Math.floor((max - min) * Math.random()) + min;
